server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pqcrypto==0.3.4
cryptography==43.0.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.6.1
httpx==0.27.2
pytest==8.2.2

